Remove random number generator
The pseudo random number generator (PRNG) might behave differently
depending on the platform and implemention of the STL. The PRNG may
thread-safe, for example, which would serialise some accesses. It may
slow down the whole code a lot as well giving the impression that the
memory allocation is slow.

The PRNG is replaced by an arithmetic expression that contains a
volatile variable. This, hopefully, prevents the compiler to do any
crazy optimisation.